+++ src/Security/Security/Models/Pricings/PSSecurityPricing.cs
@@ -25,5 +25,7 @@ namespace Microsoft.Azure.Commands.Security.Models.Pricings
         public string PricingTier { get; set; }
 
         public TimeSpan? FreeTrialRemainingTime { get; internal set; }
+
+        public string SubPlan { get; set; }
     }
 }