+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,8 @@
 
 -->
 ## Upcoming Release
+* Added breaking change message for `New-AzVmss`.
+* Added `-PerformancePlus` parameter to `New-AzDiskConfig`
 
 ## Version 5.4.0
 * Added `-SkipIdentity`, `-PathUserIdentity`, `-IsTest` parameter to `Set-AzVMAEMExtension`